admin_panel/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.6.0
django-split-settings==1.3.2
flake8==6.1.0
mccabe==0.7.0
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.1
sqlparse==0.5.2
uwsgi==2.0.20